Move bvec_bval_tool from utils to dwi module
