+++ scripts/scil_compute_msmt_frf.py
@@ -40,7 +40,7 @@ from scilpy.io.utils import (add_force_b0_arg,
                              assert_inputs_exist, assert_outputs_exist,
                              assert_roi_radii_format)
 from scilpy.reconst.frf import compute_msmt_frf
-from scilpy.utils.bvec_bval_tools import extract_dwi_shell
+from scilpy.dwi.bvec_bval_tools import extract_dwi_shell
 
 
 def buildArgsParser():
