+++ scripts/scil_screenshot_dti.py
@@ -22,7 +22,7 @@ import numpy as np
 from scilpy.io.utils import (add_overwrite_arg,
                              assert_inputs_exist,
                              assert_outputs_exist)
-from scilpy.utils.bvec_bval_tools import normalize_bvecs, get_shell_indices
+from scilpy.dwi.bvec_bval_tools import normalize_bvecs, get_shell_indices
 from scilpy.image.volume_operations import register_image
 from scilpy.viz.screenshot import display_slices
